+++ changelog.md
@@ -4,6 +4,8 @@
 
 ## Standard library additions and changes
 
+- Make custom op in macros.quote work for all statements.
+
 - On Windows the SSL library now checks for valid certificates.
   It uses the `cacert.pem` file for this purpose which was extracted
   from `https://curl.se/ca/cacert.pem`. Besides
@@ -253,6 +255,8 @@ provided by the operating system.
 - Added `then`, `catch` to `asyncjs`, for now hidden behind `-d:nimExperimentalAsyncjsThen`.
 
 - `--newruntime` and `--refchecks` are deprecated.
+
+- Added `unsafeIsolate` and `extract` to `std/isolation`.
 
 ## Tool changes